requirements.txt
# ----------------------------------------------------------
# Author: Nandan Kumar
# Assignment 13: Project Dependencies
# File: requirements.txt
# ----------------------------------------------------------

# ----------------------------------------------------------
# 1. Core Framework
# ----------------------------------------------------------
fastapi==0.115.4
uvicorn[standard]==0.32.0
Jinja2==3.1.4

# ----------------------------------------------------------
# 2. Validation and Typing
# ----------------------------------------------------------
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.7.1
annotated-types==0.7.0
typing_extensions==4.12.2
email_validator==2.2.0

# ----------------------------------------------------------
# 3. Database and ORM
# ----------------------------------------------------------
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
greenlet==3.1.1
python-dotenv==1.0.1

# ----------------------------------------------------------
# 4. Security and Authentication
# ----------------------------------------------------------
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
PyJWT==2.9.0

# ----------------------------------------------------------
# 5. HTTP and Networking
# ----------------------------------------------------------
httpx==0.27.2
requests==2.32.3

# ----------------------------------------------------------
# 6. Testing and Coverage
# ----------------------------------------------------------
pytest==8.3.3
pytest-cov==6.0.0
coverage==7.6.4
Faker==33.3.0

# ----------------------------------------------------------
# 7. End-to-End Testing
# ----------------------------------------------------------
playwright==1.48.0
pyee==12.0.0
